Create README.md
C programming was developed by Dennis Ritchie in the 1970s, is a versatile and influential language that has been a foundational tool for software development. It is a procedural language that focuses on functions and structured programming, combining high-level language features with low-level memory manipulation. 

C programs are fast and have a small memory footprint, making them suitable for system programming. C's syntax is simple and easy to learn, and it provides a range of data types, functions, and pointers for direct memory manipulation. It comes with a standard library for input/output, memory allocation, and other common operations. To write and run C programs, a C compiler is required, and it can be written in text editors or IDEs like Visual Studio Code or Code::Blocks. 

C is used in various domains, including system programming, embedded systems, game development, and scientific computing. Learning C involves learning the basics of syntax, data types, control structures, and memory management.